src/requirements.txt
openai>=1.0.0
click>=8.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
urllib3>=1.26.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
matplotlib>=3.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
semantic-compressor>=2.40.0
nltk>=3.9.0
tokenizers>=0.20.0

